separate page directory created
