+++ client/src/main.jsx
@@ -1,13 +1,13 @@
 import ReactDOM from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import App from './App.jsx';
-import Books from './Components/Books/Books.jsx';
+import Books from './Pages/Books/Books.jsx';
 import Home from './Pages/Home/Home.jsx';
-import Login from './Components/Login/Login.jsx';
-import PipeCigar from './Components/PipeCigar/PipeCigar.jsx';
-import Tabletop from './Components/Tabletop/Tabletop.jsx';
-import Whiskey from './Components/Whiskey/Whiskey.jsx';
-import SignUp from './Components/Login/SignUp.jsx';
+import Login from './Pages/Login/Login.jsx';
+import PipeCigar from './Pages/PipeCigar/PipeCigar.jsx';
+import Tabletop from './Pages/Tabletop/Tabletop.jsx';
+import Whiskey from './Pages/Whiskey/Whiskey.jsx';
+import SignUp from './Pages/Login/SignUp.jsx';
 import './index.css';
 
 const router = createBrowserRouter([
